INGK-1284 Refactor Windows pcap capture into a dedicated module
