+++ tests/test_ipv6_pcap_capture.py
@@ -1,0 +1,81 @@
+import sys
+
+import pytest
+
+if sys.platform == "win32":
+    from ingenialink.utils.ipv6_pcap_capture import PcapCapture
+
+
+pytestmark = pytest.mark.skipif(sys.platform != "win32", reason="Windows-only pcap capture")
+
+
+def test_pcap_capture_applies_ipv6_filter(mocker):
+    """Configure an Ethernet capture with the IPv6 and VLAN filter."""
+    capture = mocker.MagicMock()
+    capture.datalink.return_value = 1
+    cypcap = mocker.MagicMock()
+    cypcap.create.return_value = capture
+    cypcap.DatalinkType.EN10MB = 1
+    mocker.patch("ingenialink.utils.ipv6_pcap_capture.cypcap", cypcap)
+
+    pcap_capture = PcapCapture(r"\Device\NPF_{DEADC0FF-EEEE-4444-8888-2BF6900CBFA0}")
+
+    cypcap.create.assert_called_once_with(r"\Device\NPF_{DEADC0FF-EEEE-4444-8888-2BF6900CBFA0}")
+    capture.set_snaplen.assert_called_once_with(65_535)
+    capture.set_promisc.assert_called_once_with(False)
+    capture.set_timeout.assert_called_once_with(0.01)
+    capture.activate.assert_called_once_with()
+    capture.setfilter.assert_called_once_with("ip6 or (vlan and ip6)")
+    pcap_capture.close()
+
+
+def test_pcap_capture_converts_library_errors(mocker):
+    """Convert cypcap configuration errors into OS errors."""
+    cypcap = mocker.MagicMock()
+    cypcap.Error = RuntimeError
+    cypcap.create.side_effect = RuntimeError("unable to open capture")
+    mocker.patch("ingenialink.utils.ipv6_pcap_capture.cypcap", cypcap)
+
+    with pytest.raises(OSError, match="unable to open capture"):
+        PcapCapture("Ethernet")
+
+
+def test_pcap_capture_reports_unexpected_termination(mocker):
+    """Convert pcap's termination status into a descriptive OS error."""
+    capture = PcapCapture.__new__(PcapCapture)
+    capture._capture = mocker.MagicMock()
+    capture._capture.__next__.side_effect = StopIteration
+    cypcap = mocker.MagicMock()
+    cypcap.Error = RuntimeError
+    mocker.patch("ingenialink.utils.ipv6_pcap_capture.cypcap", cypcap)
+
+    with pytest.raises(OSError, match="terminated unexpectedly"):
+        capture.read_packet()
+
+
+def test_pcap_capture_reports_read_errors(mocker):
+    """Expose the error returned by pcap when packet capture fails."""
+    capture = PcapCapture.__new__(PcapCapture)
+    capture._capture = mocker.MagicMock()
+    capture._capture.__next__.side_effect = RuntimeError("read failure")
+    cypcap = mocker.MagicMock()
+    cypcap.Error = RuntimeError
+    mocker.patch("ingenialink.utils.ipv6_pcap_capture.cypcap", cypcap)
+
+    with pytest.raises(OSError, match="read failure"):
+        capture.read_packet()
+
+
+def test_pcap_capture_close_is_idempotent(mocker):
+    """Close an active pcap handle no more than once."""
+    capture = PcapCapture.__new__(PcapCapture)
+    pcap = mocker.Mock()
+    capture._capture = pcap
+    cypcap = mocker.MagicMock()
+    cypcap.Error = RuntimeError
+    mocker.patch("ingenialink.utils.ipv6_pcap_capture.cypcap", cypcap)
+
+    capture.close()
+    capture.close()
+
+    pcap.close.assert_called_once_with()
